requirements.txt
beautifulsoup4==4.12.3
cloudscraper==1.2.71
fastapi==0.115.6
httpx==0.26.0
pydantic==2.10.4
python-dotenv==1.0.1
slowapi==0.1.9
uvicorn==0.34.0

